Decouple Pool and MacroServer start from CMD
Since sardana have moved to setuptools the Pool and MacroServer scripts 
are not available until the moment of installation. Change concept of
this image, so by default it starts only the mysql and Tango DB.
The Pool and MacroServer may still be started with the supervisor, but
only after the sardana installation. Add documentation to reflect that.